商品削除機能の修正1

+++ app/controllers/items_controller.rb
@@ -36,8 +36,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user == item.user
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
